**6.1.0**
* Bug fix installing efi tools: typo in destination directory name.
* Improve build options
  - Compiler flag *-march=* now set by a build option: -Dcpu_level=x86-64-v3 (default is -march=x86-64)
  - Additional compiler/linker hardening
  - Release build now defaults to -Dhardened_export=true which uses -fvisibility=hidden
    Also speeds up shared lib load times.
  - Confirm project builds without any warnings using *-fanalyzer*
* efi-tool (always uses bls layout) - simplify the bls initialization.

+++ src/lib/efi_var.c
@@ -72,11 +72,10 @@ char *read_efi_var_string(const char *efi_path) {
         return nullptr;
     }
 
-    size_t f_size = st_buf.st_size;
+    size_t f_size = (size_t)st_buf.st_size;
     if (f_size <= 4) {
         return nullptr;
     }
-    //f_size -= HDR_LEN;
 
     /*
      * - bound check since efi_path is not under our control
